requirements.txt
altair==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.8
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.1
black==22.6.0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
executing==0.10.0
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
imageio==2.10.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.3.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.5.3
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.12
numpy==1.23.2
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
Pillow==9.2.0
pkg_resources==0.0.0
platformdirs==2.5.2
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
Pympler==1.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pyzmq==22.3.0
requests==2.26.0
rich==12.5.1
scikit-image==0.18.3
scipy==1.9.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.4.0
streamlit==1.12.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.11.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.3.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.8.1
